Added new page.
I also had to convert the links to not use a trailing slash. Sadly,
this causes problems locally. I hope to look into this further, it
seems like a standing issue with next.js.

What this means though is that on github, it will redirect to not
https. This doesn't really matter for this application, but it is a
bit ugly. I will hopefully find a fix at some point.